+++ CHANGELOG.md
@@ -20,6 +20,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated dependencies
 - Updated `User` class so admin users inherit all existing permissions for every tenant
 
+### Removed
+
+- Removed `SoftDeletes` trait from the following resource models:
+  - `PermissionsModel`
+  - `TenantInvitationsModel`
+  - `TenantMetaModel`
+  - `TenantRolesModel`
+  - `TenantTeamsModel`
+  - `TenantUserMetaModel`
+  - `UserKeysModel`
+  - `UserMetaModel`
+
 ## [1.0.0] - 2025.01.09
 
 ### Added